+++ praw/models/inbox.py
@@ -3,6 +3,7 @@ from ..const import API_PATH
 from .listing.generator import ListingGenerator
 from .base import PRAWBase
 from .util import stream_generator
+from .reddit.message import Message
 
 
 class Inbox(PRAWBase):
@@ -24,6 +25,47 @@ class Inbox(PRAWBase):
         """
         return ListingGenerator(self._reddit, API_PATH['inbox'],
                                 **generator_kwargs)
+
+    def collapse(self, items):
+        """Mark an inbox message as collapsed.
+
+        :param items: A list containing instances of :class:`.Message` or
+            list of t4 message fullnames.
+
+        Requests are batched at 25 items (reddit limit).
+
+        For example, to collapse all unread Messages, try:
+
+        .. code:: python
+
+            from praw.models import Message
+            unread_messages = []
+            for item in reddit.inbox.unread(limit=None):
+                if isinstance(item, Message):
+                    unread_messages.append(item)
+            reddit.inbox.collapse(unread_messages)
+
+        To collapse a list of modmail messages by fullnames, try:
+
+        .. code:: python
+
+            messages = []
+            for message in reddit.subreddit('mod').mod.inbox(limit=None):
+                messages.append(item.fullname)
+            reddit.inbox.collapse(messages)
+
+        .. seealso::
+
+           :meth:`.Message.uncollapse`
+
+        """
+        while items:
+            if isinstance(items[0], Message):
+                data = {'id': ','.join(x.fullname for x in items[:25])}
+            else:
+                data = {'id': ','.join(items[:25])}
+            self._reddit.post(API_PATH['collapse'], data=data)
+            items = items[25:]
 
     def comment_replies(self, **generator_kwargs):
         """Return a ListingGenerator for comment replies.
@@ -207,6 +249,47 @@ class Inbox(PRAWBase):
         """
         return ListingGenerator(self._reddit, API_PATH['submission_replies'],
                                 **generator_kwargs)
+
+    def uncollapse(self, items):
+        """Mark an inbox message as uncollapsed.
+
+        :param items: A list containing instances of :class:`.Message` or
+            list of t4 message fullnames.
+
+        Requests are batched at 25 items (reddit limit).
+
+        For example, to uncollapse all unread Messages, try:
+
+        .. code:: python
+
+            from praw.models import Message
+            unread_messages = []
+            for item in reddit.inbox.unread(limit=None):
+                if isinstance(item, Message):
+                    unread_messages.append(item)
+            reddit.inbox.uncollapse(unread_messages)
+
+        To uncollapse a list of modmail messages by fullnames, try:
+
+        .. code:: python
+
+            messages = []
+            for message in reddit.subreddit('mod').mod.inbox(limit=None):
+                messages.append(item.fullname)
+            reddit.inbox.uncollapse(messages)
+
+        .. seealso::
+
+           :meth:`.Message.collapse`
+
+        """
+        while items:
+            if isinstance(items[0], Message):
+                data = {'id': ','.join(x.fullname for x in items[:25])}
+            else:
+                data = {'id': ','.join(items[:25])}
+            self._reddit.post(API_PATH['uncollapse'], data=data)
+            items = items[25:]
 
     def unread(self, mark_read=False, **generator_kwargs):
         """Return a ListingGenerator for unread comments and messages.